backend/requirements.txt
asgiref==3.8.1
Django==5.1
django-cors-headers==4.4.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
Faker==28.0.0
inflection==0.5.1
Markdown==3.6
numpy==1.26.4
packaging==24.1
pyarrow==17.0.0
PyJWT==2.9.0
pylance==0.16.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
uritemplate==4.1.1
